Adds Full-Text Search support for Android.
The 1.x versions of CBLite do not support full-text search on
Android. This commit introduces a Sqlite3 database as a side-car
to use as an index for full-text search. The FTS side-car database
is updated when a change is reported by CBLite. This hook uses the
CBLite addChangeListener functionality to learn when new attendees
are inserted.

This is not production quality code. This is a hack to test the
viability of supporting full-text search on Android with the 1.x
version of CBLite. It does not handle changes other than added
records and is specific to the attendee entity. There are large
bugs that prevent this from being production ready without a
significant development investment.